// src/model/ActionMasterEnum.ts
export enum ActionMasterEnum {
  MOBILE_KEYPAD = "MOBILE_KEYPAD",
  SHARE_AUDIO = "SHARE_AUDIO",
  STOP_AUDIO = "STOP_AUDIO",
  SEND = "SEND",
  START_AUDIO = "START_AUDIO",
  SHARE_LEAD = "SHARE_LEAD",
  PLAY_VIDEO = "PLAY_VIDEO",
  SHARE_VIDEO = "SHARE_VIDEO",
  CLICK_IMAGES = "CLICK_IMAGES",
  VIEW_IMAGES = "VIEW_IMAGES",
  REVIEW_SMS = "REVIEW_SMS",
  VALUE_DATA = "VALUE_DATA",
  SHARE_ADDITIONAL_DOCUMENT = "SHARE_ADDITIONAL_DOCUMENT",
}